+++ CHANGELOG.rst
@@ -3,6 +3,9 @@
 Changelog
 =========
 
+- Prevent access to object paths containing ``__globals__`` or ``__builtins__`` in ``invoke(). Attempting to access these keys will raise a ``KeyError``.
+
+
 v8.0.5 (2025-01-17)
 -------------------